requirements.txt
aiohttp>=3.9.0
alembic==1.14.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
greenlet==3.3.1
idna==3.10
playwright==1.57.0
psycopg2-binary==2.9.11
pyee==13.0.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
soupsieve==2.6
sqlalchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.3.0
